+++ income-calculator/pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>com.techelevator</groupId>
-	<artifactId>portfolio-project</artifactId>
+	<artifactId>income-calculator</artifactId>
 	<version>1.0</version>
 	<packaging>jar</packaging>
