+++ income-calculator/src/main/resources/placeholder.txt
