+++ income-calculator/src/test/java/com/techelevator/ApplicationTests.java
@@ -1,0 +1,10 @@
+package com.techelevator;
+
+import org.junit.*;
+import org.junit.runners.MethodSorters;
+import static org.junit.Assert.*;
+
+public class ApplicationTests {
+
+}
+
