+++ income-calculator/src/test/resources/placeholder.txt
