--- wage-calculator/src/test/java/com/techelevator/ApplicationTests.java
@@ -1,26 +1,0 @@
-package com.techelevator;
-
-import org.junit.*;
-import org.junit.runners.MethodSorters;
-import static org.junit.Assert.*;
-
-public class ApplicationTests {
-
-    @Test
-    public void twoPlusTwoEqualsFour() {
-        assertEquals(4, 2 + 2);
-    }
-
-    @Test
-    public void sevenDividedByTwoIsThree() {
-        int result = 7 / 2;
-        assertEquals(3, result);
-    }
-
-    @Test
-    public void sevenDividedByTwoIsThreeAndOneHalf() {
-        double result = 7.0 / 2;
-        assertEquals(3.5, result, 0.01);
-    }
-
-}
